feat: 新しいViewクラス（HealthView、MetricsView、StocksView、VersionView）を追加し、アプリケーションのルーティングを改善しました。
